requirements.txt
PyQt6
qtawesome
PyYAML
evdev
